Improve Sequence.shouldContainExactly(...) (#3743)
Previously, in the case of a failure, simply `toString()` on the involved sequences has been invoked to build the failure message. However, this just printed the (internal) sequence class name.

Now, the consumed and compared elements are collected in separate lists and then included in the failure message. So the failure contains all elements up the point, when a difference has been detected. This approach even works for infinite sequences.

Fixes https://github.com/kotest/kotest/issues/3742