Changing the Kotlin Gradle plugin dependncy to compileOnly (#3856)
Using it as an implementation dependency seems to add a possibly stale
version of the Kotlin plugin to the build path. Normally a user would
already have a possibly different version of Kotlin pulled in, and these
could clash (as seen in issue #3850)

Could this cause regressions that I'm not thinking of?