Removing Arb.create which seems to not exist anymore
I tried to figure out when it was introduced to see what happened to it.
It seems to have been around when the list of generators was first created in 2020
so it was hard to see exactly where it was added.

Just removing the function from the list for now to avoid further confusion.