Fix the snapshot version badge in index markdowns (#3888)
<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->

#3857 
Started from this issue.

Kotest migrated to a new Sonatype repository a while back which caused
the badges to break.
it changed to s01.oss.sonatype.org now.

---------

Co-authored-by: Emil Kantis <emil.kantis@protonmail.com>