fix(docs): indentation in System Extensions (#3940)
This PR updates some of the indentations in the system extensions files
in the docs. There is a mix of two space and four space indentations in
the docs, but as Kotlin seems to recommend four spaces this is what I
went for here.

No content has changed. I also didn't update all relevant versions. At
least it will look nice for the next release.


<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->