Add kotlin-faker to extensions list (#3971)
<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->

Hi,
I've been using kotest for many years now, and thought it would be nice
to contribute to the kotest community by providing an extension for
kotest property testing, which allows one to generate even a larger
variety of data than the built-in Arbs.

Should I also update other versions of the docs? Which ones?

Also maybe you want to consider adding a link to the extension somewhere
in kotest property docs? E.g.
https://kotest.io/docs/proptest/property-test-extra-arbs.html

---------

Co-authored-by: Emil Kantis <emil.kantis@protonmail.com>