Allow comparison between empty set and other iterables (#4005)
We disallow comparison between sets and other iterables when the
iteration order is not clear. For an empty set, there is nothing to
iterate so there is no such ambiguity.

Fixes #4004

<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->