Add an Arbitrary to generate whitespace. (#4030)
This pull request adds a `whitespace` function to the Codepoints in
order to generate strings with only whitespace characters (and compose
more complex generators). I happened to need one for a project and
thought I might share it.

Since this is my first contribution, I'm not sure of which docs I should
update where (and also, I've never multiplatform'd in kotlin before, so
I'm not sure if I need to do anything else than the code I've written
for this to work everwhere).

Cheers :)

Co-authored-by: Joachim Desroches <jo@thedesroches.net>