Fix test name comparison in docs (#4034)
<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->

**Thanks for creating a great testing tool called kotest🙂**

While reading the documentation, I found something wrong and fixed it.


![image](https://github.com/kotest/kotest/assets/78959175/f9a59820-adfa-4f7e-93f7-baa745be9468)

In the Conditional tests with enabled flags example code, I'm using
`it.name.startsWith(...)`, whereas in the current specification I can
access it as `it.name.testName.startsWith(...)`.

I've seen the error appear in versions 5.2 to 5.9, and I've fixed them
all.
(Please excuse the awkwardness of the PR due to translation)

I always support kotest👍