Add Stacktrace Substring Matching to Throwable Matchers (#3998)
This PR introduces the ability to match substrings and regular
expressions in the stacktrace of a `Throwable` in the Kotest assertions
library. This feature can be useful when writing tests that need to
assert on specific parts of a stacktrace.

Changes:

1. Added two new infix functions `shouldHaveStacktraceContaining` and
`shouldNotHaveStacktraceContaining` to the `Throwable` class. These
functions accept a `String` or `Regex` and use it to match against the
stacktrace of the `Throwable`.
2. Added corresponding matcher functions `haveStacktraceContaining` that
create a `Matcher<Throwable>` to perform the actual matching logic.
3. Added tests for these new functions in `ThrowableMatchersTest.kt`.

This enhancement provides more flexibility when writing tests involving
exceptions and their stacktraces.

---------

Co-authored-by: Sam <sam@sksamuel.com>
Co-authored-by: Emil Kantis <emil.kantis@protonmail.com>