Add missing annotation to actual declarations (#4165)
`@JVMOnly` is defined on the `expect fun`, but not on the `actual fun`s.

This triggers a warning:

> Annotation `@JVMOnly` is missing on actual declaration.
> All annotations from expect `fun coroutineErrorCollectorInterceptor():
TestExecutionInterceptor defined in io.kotest.engine.test.interceptors
in file interceptors.kt` must be present with the same arguments on
actual `fun coroutineErrorCollectorInterceptor():
TestExecutionInterceptor defined in io.kotest.engine.test.interceptors
in file CoroutineErrorCollectorInterceptor.kt`, otherwise they might
behave incorrectly.

Part of #4085