Cleaning stacktraces later (#4051)
When we throw an AssertionError and the matcher is actually used by an
inverted matcher (i.e. we expect a failure), then the stacktrace is
irrelevant.

Partially fixes #4041 
#4048 fixes rest, but this fix can go into next patch and fixes 2/3rds
of the performance costs

<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->

Co-authored-by: Sam <sam@sksamuel.com>