Use Java 21 to run Gradle Daemon (#4221)
To maximise Gradle cache hits, and to use the best Gradle features, use
Java 21 to run Gradle builds.

Gradle has a new feature for running the Gradle Daemon with a fixed Java
version. This can be configured to run automatically. Currently Gradle
does not auto-provision the necessary Java version, but hopefully this
will be available in Gradle 8.10.

This PR should be merged after #4218, which will ensure that a stable
JDK is used for compilation and running tests.

---------

Co-authored-by: Sam <sam@sksamuel.com>