Cache Kotlin Konan directory
Kotlin uses a persistent directory for caching dependencies and transforms. (The directory is `~/.konan` by default, or can be set manually using `KONAN_DATA_DIR`).

On CI the directory is not persisted, so Kotlin will re-do the same download and transformations. Caching the directory will improve performance, and is now a viable option because the GitHub Actions Cache is less overloaded (thanks to Remote build cache and fixes in the Gradle GitHub Action.)