Submatching for collections, part 1 (#4244)
Start matching sub-collections when the whole thing does not match. The
plan is to use this approach for the following cases:

* startWith
* endWith
* include/contain
* containExactly

I'd like to use this approach for Collection, String, and Sequence.

Find some longest sub-collections that match, For instance, if we assert
that

`"I'm not coming this winter" shouldInclude "winter is coming"`

 - this algorithm finds "coming" and "winter".