StringPartialMatches.kt  (#4335)
add more detailed description when `shouldInclude` fails, but we can
find a substring than matches, such as:

```kotlin
         "should find a submatch for reasonably long value and substring" {
            val line = "The quick brown fox jumps over the lazy dog"
            shouldThrow<AssertionError> {
               line shouldInclude "The coyote jumps over the lazy dog"
            }.message.shouldContainInOrder(
               """"The quick brown fox jumps over the lazy dog" should include substring "The coyote jumps over the lazy dog"""",
               """Match[0]: expected[10..33] matched actual[19..42]""",
               """Line[0] ="The quick brown fox jumps over the lazy dog"""",
               """Match[0]= -------------------++++++++++++++++++++++++"""
            )
         }
```

I know we have a matcher for multi-line strings, but it does not seem to
be very helpful.

The plan is to extend this approach to many other places where strings
are matched, and also look for similar strings in `similarity` package,
like how we search for similar data classes.