Modernize shouldBeUnique collection assertions (#4375)
- Adds shouldBeUnique for primitive arrays
- Prints indexes of duplicated elements
- Sequence.shouldBeUnique delegates to the Iterable matcher

As before, the removal of any leftover Collection overloads will be
performed in a followup PR.

#4369 introduced a helper class `DuplicationReport` for finding indexed
duplicates in a container using `equals`. This helper can be used for
implementing `shouldBeUnique` as well. `shouldBeUnique` offers two
variants: checking for uniqueness by equals and by using a comparator.
Therefore, `DuplicationReport` was renamed to
`DuplicationByEqualsReport` and a sibling helper,
`DuplicationByCompareReport`, was introduced.

---------

Co-authored-by: Sam <sam@sksamuel.com>