Change edgecase(rs): A? to return Sample<A>? (#4289)
This would address https://github.com/kotest/kotest/issues/4029 and make
return types of edgecase and sample consistent.

Caveat: it requires any edgecaseFn (e.g., in arbitrary calls) to define
a function that returns a Sample<A> instead of A - that's not really
nice. However, it's still better than the alternative: still allowing a
special meaning of "null" in those functions, as that would be
inconsistent and probably confusing.

Keeping non-nullable alternatives wouldn't work nicely due to name
clashes.

It might be worth considering other alternatives, but I can't think of
any that hasn't any major drawbacks.
Maybe, instead using Sample<A>, introduce a new interface Edgecase<A>?

---------

Co-authored-by: Sam <sam@sksamuel.com>
Co-authored-by: Alex Kuznetsov <Alex.Cue.987@gmail.com>