Add JSON Matchers For `File`/`Path` (#4461)
Adds JSON matchers that read the text content of a `File`/`Path`

I wanted matchers that did this for my project and decided to contribute
them to kotest.
Since `File` and `Path` are jvm-only, I added them to the `jvmMain`
sourceset.

However, adding [kotlinx-io](https://github.com/Kotlin/kotlinx-io) as a
compile only dependency to the json matchers module so that anyone who
has that dependency can also have some matchers for that.
I didn't do this, as I wasn't sure if adding a dependency is desired.
However if it is, I could add it.

---------

Co-authored-by: Alex Kuznetsov <Alex.Cue.987@gmail.com>