feat: add Yaml matchers module (#4443)
Hi everyone,
I started working on the Yaml matchers feature. In order to have a
starting point for discussions/suggestions/improvements I open this PR
quite early in draft mode. I had to set the Java version to 11 for now
since I'm still working on upgrading KAML to Java 8 compatibility.
The 
```
   id("kotest-jvm-conventions")
   id("kotest-js-conventions")
```
can also be replaced later with the multiplatform-conversions, but this
also requires a newer KAML version.
If you have ideas for additional matchers or a different file structure
feel free to point it out, I'll happily adjust the code.

---------

Co-authored-by: Ronny Bräunlich <ronny.braeunlich@cbc.de>