Added strawman proposal for next gen property test syntax (#4258)
Following on from inspiration by @BenWoodworth in
https://github.com/kotest/kotest/issues/3445 I have been playing around
with the following syntax as a potential for Kotest 6.0 (now that we're
deep in the work for that major release :))

Goals:
1. To make property testing easier to read and write by taking advantage
of Kotlin's language features.
2. To support new features by avoiding legacy syntax.
3. To enhance property testing support by coupling together with kotest
framework.

To point 1, after many years of development in Kotest's property test
framework, we've added a lot of features that can seem more intuitive
and easier to understand if we took a fresh look based on what we now
know. For instance, the way we configure config is a bolt on to the
original syntax, and the way arbs are passed into `checkAll` and
`forAll` can seem clunky.

To point 2, I am not trying to make all this work with the existing
code. The main interfaces - `Gen`, `Arb`, `Exhaustive`, `Classifier`,
and `Shrinker` - are untouched but I have avoided trying to make the
`checkAll` and `forAll` methods backwards compatible.

To point 3, it is not a goal to make this work outside of Kotest. The
original prop test library is fully compatible with JUnit and any other
framework - and will always remain so - but this alternative would be
tied into Kotest. This would allow us to have nicer syntax for defining
tests directly in specs, as well as coupling into callbacks.

Here is what I have working in a very ugly branch. If we like this work,
I would need to clean it all up, port over the rest of the features and
tests. However, now that I have property testing iterations and
shrinking working, I don't think there's anything left that would pose a
problem to migrate.

Things to spot:
1. Property tests are defined directly inside the spec, no need for a
"test" wrapper, etc. A prop test can be placed inside any container (eg
describe in DescribeSpec) or simply at the top level
2. Config values are defined inline. No need to pass in PropConfig. You
can reuse a config object though if you wish, similar to #4251
3. Arbs are specified outside the property block, but we can switch
that. I have no strong preference and it does not matter to the way it
works.
4. The actual testing iteration takes place inside a `property` block
inside the test
5. Notice the callbacks in the final test.

```kotlin
class DelegateTest : FunSpec() {
   init {
      proptest("delegate example 1") {

         iterations = 42 // run this property test 42 times
         failOnSeed = true // will fail if a seed is set
         edgecaseFactor = 0.1 // 10% of generated values will be edgecases
         shouldPrintGeneratedValues = true // output each generated value

         val a by gen { Arb.int() }
         val b by gen { Arb.int() }

         property {
            a + b shouldBe a + b
         }
      }

      proptest("delegate example 2") {

         duration = 100.milliseconds // run this property test for as many iteration as 100ms allows
         minSuccess = 5 // require 5 successful tests to ensure we get at least some in a duration based test
         maxFailure = 1 // allow 1 failure for a flakey test
         shrinkingMode = ShrinkingMode.Off // disable shrinking on this prop test

         val first by gen { Arb.string() }
         val second by gen { Arb.string() }

         property {
            val concat = first + second

            concat shouldStartWith first
            concat shouldEndWith second
            concat shouldHaveLength (first.length + second.length)
         }
      }

      proptest("shrinking") {

         seed = 838127382173 // override seed
         shouldPrintShrinkSteps = true // print shrink steps

         val a by gen { Arb.int() }
         val b by gen { Arb.int() }

         property {
            a * b shouldBe a + b
         }
      }

      proptest("assumptions") {

         val a by gen { Arb.int() }
         val b by gen { Arb.int() }
         val c by gen { Arb.int() }

         assume { a != b }
         assume { a != c }
         assume { b != c }

         property {
            a + b + c shouldBe c + b + a
         }
      }

      proptest("callbacks") {

         val a by gen { Arb.int() }
         val b by gen { Arb.int() }

         beforeProperty {
            // setup code
         }

         afterProperty {
            // some tear down code
         }

         property {
            a + b shouldBe b + a
         }
      }
   }
}
```

Would be interested in general thoughts on if this is something worth
perusing, as well as feedback on what I might have missed that might
prove difficult or impossible to do (and thus nix the whole idea).