Drop Nulls Collection Matchers (#4613)
This commit removes the redundant functions for Collections. These
functions are already covered by Iterable, as Collection is a subtype of
 Iterable.

This change does binary changes, but we don't need to deprecate. Users
won't need to make any changes to their tests, as the compiler will
simply pick up the new function signatures.

Signed-off-by: Leonardo Colman Lopes <dev@leonardo.colman.com.br>