Rework isolation modes and concurrency (#4626)
Work in progress. As we move closer to 6.0, this PR attempts to bring in
a permanent solution for concurrency in spec and tests, as well as
simplify the test engine.

* Removes isolation per leaf and isolation per test. These isolation
modes were added before coroutines even existed in Kotlin! This will be
a large breaking change, but one that hugely simplifies the test engine,
and removes lots of awkward edge cases around instantiation of specs
that then require tests to be replayed. This will make the unified
JS/JVM/Native test runner easier to accomplish. And it will allow us to
make coroutine logic more sane in how tests inherit coroutine scopes
etc.
* Removes the `SpecRunner` instances and squash these into a single
SpecExecutor2 (odd name, but that's because in the next follow up PR,
the two layers of spec exectors will be merged).
* Adds a new `IsolationPerRoot` option to replace the above that gives
you a fresh spec per top level test, which avoids replaying as root
tests are registered but not executed at spec creation.
* Removes all previous parallelism settings in specs and project config
such as `threads`, `parallelism`, `affinity`, `testConcurrency` and
`specConcurency`. Other than `threads`, these settings have been marked
as experimental since they were added.
* Adds new `TestExecutionMode` which sets concurrency inside a spec and
works for both single instance and instance per root isolation modes.
When set to `Concurrent` all root tests are launched concurrency. When
left to the default `Sequential` each root test is launched when the
previous completes. Root tests are always launched into their own
coroutine regardless and this setting just determines how many are
launched in parallel.
* Adds new `SpecExecutionMode` which sets concurrency across the engine
for specs. When set to `Concurrent` all specs tests are launched
concurrency. When left to the default `Sequential` each spec is launched
when the previous spec completes.
* The `CoroutineDispatcherFactory` usage has been simplified. Instead of
passing around a default (which was a no-op) the default has been
removed, and the engine just doesn't apply one if there isn't one set on
the spec / project configuration. This means we don't have to weave a
factory through the interceptors and scopes for no reason.
* The affinity option has been removed. It's simple enough to use a
custom factory (and we provide a new
`ThreadPerSpecCoroutineContextFactory` to do this) if the option is
needed, but it's probably not even something you want to do.

Fixes #3311
Closes #3962

---------

Co-authored-by: Leonardo Colman Lopes <dev@leonardo.colman.com.br>