Add system property key for displayFullTestPath (#4643)
<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->

Added another system property key,
`kotest.framework.testname.display.full.path`, used to configure the
option to display the full test paths in reports. This is a nice-to-have
for my use case where I want to upload my test results for CI and have
to use this setting, but currently I can only use an
`AbstractProjectConfig`.