Nullable receivers for Boolean matchers (#4807)
<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->
Nullable Booleans are possible, and must go through value checks just
like nonnull Booleans. While it's tempting to reach for the elvis
operator (`?:`), the equality operator is preferred.

This is the inspiration behind allowing the Boolean matchers to have
nullable receivers. For testing purposes, it's simply more convenient.

---------

Co-authored-by: Sam <sam@sksamuel.com>