make-eventually-behave-consistently (#4841)
Fix https://github.com/kotest/kotest/issues/4838

Before the fix, `expectedExceptions = setOf` did not work as documented.
Instead, it was retrying on every exception. But if the same exceptions
were checked for in a function, it did work.

This fix makes `expectedExceptions = setOf` work exactly the same as the
function checking for the same exceptions. To make that happen, I had to
change the requirements - `expectedExceptions` no longer supplements the
function. Instead, it overrides the function.

With the fix, `eventually` work consistently in both following cases:

when we define a Set:

```kotlin
val config = eventuallyConfig {
            duration = 5.seconds
            expectedExceptions = setOf(IOException::class)
         }
```

and when we define a function:

```koltin
val config = eventuallyConfig {
            duration = 5.seconds
            expectedExceptionsFn = { ex -> ex is IOException }
         }
```