Fix `before` and `after` matchers error messages (#4848)
The matchers were correct, but the inverted error messages mistakenly
swapped the `value` and `anotherInstant` variables, leading to incorrect
error messages.

I added simple tests asserting the correct messages for the `before` and
`after` matchers.

I would also suggest to align the wording and punctuation (which I did
not do as part of this bug fix - please fix or open an issue in case you
agree):
- "but it's not" is not consistent with the other formulations and
should probably be removed - it is clear from the context (failing
assertion)
- Some error messages end with a ".", some do not

This should probably be merged both into master and 5.9 if a 5.9.2
release should be made. I based the branch on 5.9.1.

---------

Co-authored-by: Alex Kuznetsov <Alex.Cue.987@gmail.com>