Support reflective binding of java classes (#4907)
Fixes #4901 with minimal changes.

We can consider adding support for customizing exactly which constructor
should be used if there's enough desire for it. I played around with
doing so in
https://github.com/kotest/kotest/commit/5e6cea232f12eb0a2fd5fab7aa76532ca248d267,
but I feel the cost of adding that support is not warranted at this
point in time.

If we get reasons to revisit the jvmbind APIs for some other reason, I
think it would be a good thing to consider including at that time.

<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->

---------

Co-authored-by: Alex Kuznetsov <Alex.Cue.987@gmail.com>