Add KSP processor to replace compiler plugin (#4913)
If we move to KSP we can use a stable and supported API and not the
experimental compiler plugins. This means we shouldn't have to keep
making changes each time a new version of Kotlin is released.

This implementation brings full nested test support to JS and Native. 

Also individual tests are now working from intellij for JS platforms, so
can you click the run button and run a single test. I still need to
confirm that we can do the same for Kotlin Native.


![image](https://github.com/user-attachments/assets/e6b9ba86-3d11-4902-af6b-0fa7ffc91c1f)

There are some further changes required to robust this up that I'll do
in follow up PRs.