Enable coroutineTestScope on js and wasmJs (#4813)
<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->

This is a very simple change and I am not deeply in the kotest codebase.
I used this workaround for some projects internally and it seems to
work.

If these changes are acceptable I'd be able to add some tests for this
too, akin to the already existing tests for the coroutineTestScope on
the jvm. As I am not too familiar it would be a great help if you could
point me to the right direction for adding tests on js and wasmjs.

This might close #4077

Co-authored-by: Sam <sam@sksamuel.com>