Adding signing secrets to gradle plugin workflow
This shouldn't really be needed, as the release version of the Gradle plugin is
published to Gradle Plugin Portal. It'll do as a workaround for now.

The root problem is probably caused by the Gradle plugin having a Maven publication in the first place, but that is convenient when testing the plugin locally etc.