Cleaning up some logging and unnecessary dependency wiring (#4941)
I made an attempt to add a dependency to the generateMavenPom task in
the hope that it would let the maven publication get created before the
isPublicationEnabled property was resolved. This didn't work, but
removing the cached input seem to have been enough.

<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->