Fix `InstancePerLeaf` execution order (#5044)
<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->

ref https://github.com/kotest/kotest/issues/5022

first, let me appreciate your quick fix for my issue.

unfortunately, your fix seemed not to fix completely
https://github.com/kotest/kotest/issues/5022#issuecomment-3219284408

i tried to fix it by myself. please kindly review my first PR
___

## Solution

the same `LeafLaunchingScope` instances are executed in siblings of test
nodes (Container or Leaf).
it means that the first node is always fresh (or seed), and the second
and latter nodes always should be executed in fresh spec.
in my solution, i added `hasVisitedFirstNode` property to
`LeafLaunchingScope` to manage this.
if the visited nodes are not first one, the node and children nodes will
be executed in a fresh spec.
it means that container nodes also can be treated as `target`. so, i
extended `executeInFreshSpec` and `target` object comparison

## Concerns

- i removed `AtomicBoolean` and use just `Boolean` instead because the
property is enclosed in `LeafLaunchingScope`. do you think this is ok?
- i execute Container type in `executeInFreshSpec`. to do that, i remove
`require` block. is that no problem?
- i added some comments and logger. if you think there are more
appropriate messages, please let me know.

thank you

---------

Co-authored-by: Sam <sam@sksamuel.com>
Co-authored-by: Alex Kuznetsov <Alex.Cue.987@gmail.com>