Add missing Extension marker to core extension and listeners. (#5097)
Fixed a bug where the `InstantiationListener` was not marked as an
`Extension` like the other listeners are.
It could not be registered directly, or found in the extension registry
[here](https://github.com/kotest/kotest/blob/090e32115e0e82763ec1d6bbdeb81566a5530039/kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/spec/SpecExtensions.kt#L92),
unless the implementing class was marked as an `Extension` separately by
the user:

```kotlin
suspend fun specInstantiated(spec: Spec) = runCatching {
    logger.log { Pair(spec::class.bestName(), "specInstantiated $spec") }
    specConfigResolver.extensions(spec)
        .filterIsInstance<InstantiationListener>()
        .forEach { it.specInstantiated(spec) }
}
```

Same for `SpecExecutionOrderExtension` and `SpecRefExtension`.

---------

Co-authored-by: Sam <sam@sksamuel.com>