Fix kotest plugin in project with Kotlin 2.2.20 (#5106)
Calling `target.publishable` throws an exception in a Kotlin
multiplatform project when using Kotlin 2.2.20. This seems to be a
Kotlin internal issue because `getOrThrow()` is called on a
non-completed future and I can't find an exposed API to complete it.

As a workaround instead of checking if the target is publishable I check
if the original task is enabled or not.

Co-authored-by: Sam <sam@sksamuel.com>