`shouldBeEqualToUsingFields` with different types (#5124)
<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->
#### `shouldBeEqualToUsingFields` with different types
Addressing my own feature request in
https://github.com/kotest/kotest/issues/5123

This change introduces the ability in `shouldBeEqualToUsingFields` to
allow for equality assertions between objects of different types.
This is useful in scenarios where you want to compare objects that share
common properties but are not of the same class, such as comparing a
domain model with a DTO.

#### Changes
- `KProperty1<T, *>` has been introduced instead of `KProperty<*>` to
ensure at compile time that the properties provided belong to the
caller's type `T`.
- The internal comparison logic has been updated to look up the
corresponding property by name on the other object, in case `T` and `V`
are of different class.

---------

Co-authored-by: Alex Kuznetsov <Alex.Cue.987@gmail.com>