make shouldCompile/shouldNotCompile support multiple files (#5182)
<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->

Title as is

The current implementation of kotest-assertions-compile can only pass
one file to CodeSnippet.
To handle compile testing that requires multiple files, we have added
variations of shouldCompile/shouldNotCompile that take multiple files as
receivers.

---------

Co-authored-by: Alex Kuznetsov <Alex.Cue.987@gmail.com>