Skip modules in PR on !linux when JVM only (#5199)
For modules that are JVM only, there's no need to build and test on all
runners, linux is sufficient. We don't need to test that Oracle
implemented the JVM properly on all OSes :)
This should speed up the macos and windows builds significantly.
We'll still build everything on master, this is for PRs only.