Update to Gradle 9.2.1 & remove some potentially old workarounds (#5176)
A few changes required to get this updated to 9.2.1:
* The workaround in `buildSrc/build.gradle.kts` was causing a
compilation failure. It seems to be trying to get around [this
issue](https://github.com/gradle/gradle/issues/22647), but it's a lot to
go through to just suppress a meaningless warning. Probably ok without
it.
* Enabled configuration cache, as 9.2.0 is supposed to [have big
performance
improvements](https://docs.gradle.org/9.2.0/release-notes.html#performance-improvements)
in that area & earlier problems that caused it to be disabled may no
longer exist.
* The windows build was also overriding the default build dir, which was
causing an issue with plugins, so removed that here. Doesn't seem to
have a detrimental effect.
* Updated a few of the github actions